Revised edition about the repositories
where the memory repositories architecture has been transformed into
database repositories architecture.
A DbRepositoryImpl has been created, that is compatible with type T
objects to be stored in the database, with primary key of type K. The
basic crud operations are implemented in the DbRepositoryImpl. The
DbPropertyRepairRepositoryImpl, DbPropertyRepositoryImpl,
DbUserRepositoryImpl have been created in order to be implemented the
operations that are not common for all the entities in the database,
like updates. The DbRepositoryImpl has an EntityManager whose value is
set using construction injection, that is an EntityManager object is
injected when the class in instantiated. Each of the
DbPropertyRepairRepositoryImpl, DbPropertyRepositoryImpl,
DbUserRepositoryImpl inherits the entity manager
from DbRepositoryImpl. The domain package has been renamed to model, as
it will contain the entities of the database.